+++ internal/handlers/remoteEvents/readyEvent/ready.go
@@ -6,8 +6,8 @@ import (
 	globalConfiguration "github.com/RazvanBerbece/Aztebot/internal/globals/configuration"
 	globalRepositories "github.com/RazvanBerbece/Aztebot/internal/globals/repositories"
 	channelHandlers "github.com/RazvanBerbece/Aztebot/internal/handlers/channelEvents"
-	cron "github.com/RazvanBerbece/Aztebot/internal/services/tasks/cron"
-	"github.com/RazvanBerbece/Aztebot/internal/services/tasks/startup"
+	cron "github.com/RazvanBerbece/Aztebot/internal/tasks/cron"
+	"github.com/RazvanBerbece/Aztebot/internal/tasks/startup"
 	"github.com/RazvanBerbece/Aztebot/pkg/shared/utils"
 	"github.com/bwmarrin/discordgo"
 )
