+++ internal/tasks/startup/publishDefaultEmbeds.go
